Refactor AssetAmount API structure
