+++ crates/indexer/src/api/borrowers/dto.rs
@@ -1,13 +1,7 @@
 use serde::Serialize;
 use utoipa::ToSchema;
 
-#[derive(Serialize, ToSchema)]
-pub struct AssetAmount {
-    pub asset: String,
-    /// Amount in satoshis (decimal string).
-    #[schema(example = "1000")]
-    pub amount: String,
-}
+use crate::api::dto::AssetAmount;
 
 #[derive(Serialize, ToSchema)]
 pub struct BorrowerOverview {
