+++ crates/indexer/src/api/db/mod.rs
@@ -1,0 +1,3 @@
+mod asset_sums;
+
+pub(crate) use asset_sums::{AssetSumRow, asset_amounts_from_rows};
