+++ crates/indexer/src/api/dto/asset_amount.rs
@@ -1,0 +1,10 @@
+use serde::Serialize;
+use utoipa::ToSchema;
+
+#[derive(Serialize, ToSchema)]
+pub struct AssetAmount {
+    pub asset: String,
+    /// Amount in satoshis (decimal string).
+    #[schema(example = "1000")]
+    pub amount: String,
+}
