+++ crates/indexer/src/api/dto/mod.rs
@@ -1,0 +1,3 @@
+mod asset_amount;
+
+pub use asset_amount::AssetAmount;
