+++ crates/indexer/src/api/mod.rs
@@ -1,4 +1,6 @@
 mod borrowers;
+mod db;
+mod dto;
 mod error;
 mod factories;
 mod lenders;
@@ -10,6 +12,7 @@ pub mod server;
 mod state;
 pub mod utils;
 
+pub use dto::AssetAmount;
 pub use error::*;
 pub use openapi::ApiDoc;
 pub use params::*;
